problem with newpostcard
