--- app/src/main/java/ru/netology/nmedia/activity/PostCardLayout.kt
@@ -1,36 +1,0 @@
-package ru.netology.nmedia.activity
-
-import android.content.Intent
-import android.net.Uri
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import ru.netology.nmedia.R
-import ru.netology.nmedia.databinding.ActivityPostCardLayoutBinding
-
-class PostCardLayout : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        val binding = ActivityPostCardLayoutBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
-
-        binding.videoView.setOnClickListener {
-            //получаем ссылку
-            val url = Uri.parse(binding.tvVideoPublished.toString())
-            //создаем интент
-            val intent = Intent(Intent.ACTION_VIEW, url)
-            startActivity(intent)
-        }
-    }
-
-
-}
