+++ app/src/main/java/ru/netology/nmedia/activity/PostCardLayoutFragment.kt
@@ -3,50 +3,49 @@ package ru.netology.nmedia.activity
 import android.annotation.SuppressLint
 import android.content.Intent
 import android.net.Uri
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.viewModels
-import androidx.constraintlayout.widget.Group
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import ru.netology.nmedia.Post
 import ru.netology.nmedia.R
-import ru.netology.nmedia.R.id.tv_author
-import ru.netology.nmedia.R.id.tv_content
 import ru.netology.nmedia.activity.FeedFragment.Companion.textArg
-import ru.netology.nmedia.viewmodel.PostViewModel
 import ru.netology.nmedia.adapter.OnInteractionListener
 import ru.netology.nmedia.adapter.PostAdapter
+import ru.netology.nmedia.databinding.ActivityPostCardLayoutBinding
 import ru.netology.nmedia.databinding.FragmentFeedBinding
-import ru.netology.nmedia.util.StringArg
+import ru.netology.nmedia.viewmodel.PostViewModel
 
-class FeedFragment : Fragment() {
+class PostCardLayoutFragment : Fragment() {
     val viewModel: PostViewModel by viewModels(
         ownerProducer = ::requireParentFragment
     )
-
-    companion object {
-        var Bundle.textArg: String? by StringArg
-    }
-
     override fun onCreateView(
-        inflater: LayoutInflater, //класс в android, используемый для создания view
-        container: ViewGroup?, //специальный тип в android, используемый как контейнер для других видов
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         //надуваем разметку
-        val binding = FragmentFeedBinding.inflate(
+        val binding = ActivityPostCardLayoutBinding.inflate(
             inflater,
             container,
             false
         )
 
-        //теперь имеем возможность обращаться к группе элементов
-        val groupVideo = view?.findViewById<Group>(R.id.group_video)
+        view?.let {
+            ViewCompat.setOnApplyWindowInsetsListener(it.findViewById(R.id.main)) { v, insets ->
+                val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+                v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+                insets
+            }
+        }
 
         val adapter = PostAdapter(object : OnInteractionListener {
             override fun onEdit(post: Post) {
@@ -101,52 +100,29 @@ class FeedFragment : Fragment() {
                 viewModel.openPost(post)
                 findNavController().navigate(R.id.action_feedFragment_to_postCardLayoutFragment)
             }
+
+
         })
 
 
-//        val newPostLauncher = registerForActivityResult(NewPostContract) {
-//            //получаем результат
-////            val result = it ?: return@registerForActivityResult
-////            viewModel.changeContentAndSave(result)
-//
-//            //получаем результат лаунчера
-//            val result = it
-//            //выполняем проверку на null
-//            if (it !== null) {
-//                if (result != null) {
-//                    viewModel.changeContentAndSave(result)
-//                }
-//            } else {
-//                //отменяем редактирование и очищаем пост
-//                viewModel.cancelEdit()
-//                //выходим из метода
-//                return@registerForActivityResult
-//            }
-//        }
-
-        binding.list.adapter = adapter
         viewModel.data.observe(viewLifecycleOwner) { posts ->
             val newPost = posts.size > adapter.currentList.size
             adapter.submitList(posts) {
-                if (newPost) {
-                    binding.list.smoothScrollToPosition(0) //сверху сразу будет отображаться новый пост
-                }
+//                if (newPost) {
+//                    binding.list.smoothScrollToPosition(0) //сверху сразу будет отображаться новый пост
+//                }
             } //при каждом изменении данных мы список постов записываем обновленный список постов
         }
 
-        //клик на кнопку добавить пост
-        binding.bottomSave.setOnClickListener {
-            findNavController().navigate(R.id.action_feedFragment_to_newPostFragment)
-        }
 
+        binding.videoView.setOnClickListener {
+            //получаем ссылку
+            val url = Uri.parse(binding.tvVideoPublished.toString())
+            //создаем интент
+            val intent = Intent(Intent.ACTION_VIEW, url)
+
+
+        }
         return binding.root
     }
 }
-
-
-
-
-
-
-
-
