+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -9,5 +9,6 @@ interface PostRepository {
     fun likeById(id: Long)
     fun removeById(id: Long)
     fun save(post: Post)
+    fun openPostById(id: Long) : Post
 }
